added hangman
I had originally submitted this website for my portfolio Standard #3 on Repl.it and it had the Hangman page in there BUT it was not the version using functions so I completely updated the hangman.html page and used the authors code as the layout and proceeded to add the functions per the instructions of the book, JS fo Kids. I just kept all the other pages of the website the same.