Add files via upload
Hi,

I have some issue I have been unable to resolve in the code. Right now, it asks for a letter, shows an update that does not include if you guess was correct, proceed to another pick, then reveals the answer as if the end of the game was achieved. I believe the issue lies in my var word because when I try to loop it in the updateGameState function, I receive a "var word not defined" error. Just couldn't find the issue.